added docs and comments
changed vehicles arraylist to hashmap
added security for login and a regex checker.

+++ src/vehicle_pkg/Suv.java
@@ -5,10 +5,10 @@ import enums_pkg.Rank;
 
 public abstract class Suv extends Vehicle
 {
-	private double m_Height;
+	private int m_Height;
 	private boolean m_Tow; 
 	
-	public Suv(String m_Plate,int m_EngineCC, int m_Doors, int m_Seats, int m_Laggage, String m_Model, Rank m_Rank, Gear m_Gear, double m_Height, boolean m_Tow,int m_Price) 
+	public Suv(String m_Plate,int m_EngineCC, int m_Doors, int m_Seats, int m_Laggage, String m_Model, Rank m_Rank, Gear m_Gear, int m_Height, boolean m_Tow,int m_Price) 
 	{
 		super(m_Plate,m_EngineCC,m_Doors,m_Seats,m_Laggage,m_Model,m_Rank,m_Gear,m_Price);
 		this.m_Height = m_Height;
@@ -19,10 +19,10 @@ public abstract class Suv extends Vehicle
 		super();
 	}
 
-	public final double getM_Height() {
+	public final int getM_Height() {
 		return m_Height;
 	}
-	public final void setM_Height(double m_Height) {
+	public final void setM_Height(int m_Height) {
 		this.m_Height = m_Height;
 	}
 	public final boolean isM_Tow() {
